refactor: extract GitHub data contracts
